fixes and bytesize
