+++ zea/models/preset_utils.py
@@ -13,9 +13,8 @@ from huggingface_hub.utils import EntryNotFoundError, HFValidationError
 
 import zea
 from zea.internal.cache import ZEA_CACHE_DIR
-from zea.internal.preset_utils import _hf_parse_path
+from zea.internal.preset_utils import _hf_login, _hf_parse_path
 from zea.internal.registry import model_registry
-from zea.tools.hf import _hf_login
 
 HF_PREFIX = "hf://"
